+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2025-05-24 - Vlad
 ### Updated
 - Simplify UserService
+- Simplify /messages
 
 ## [0.1.0] - 2025-05-22 - Vlad
 ### Added